[refactor] move `CSPRepLayer` to op.custom
